feat: adjust header component in storybook
