+++ src/page/ThemePage.tsx
@@ -1,7 +1,14 @@
-import Header from '@/components/Header';
+// export type DefaultHeaderProps = {
+//   label: string;
+//   title: string;
+//   description: string;
+//   backgroundColor: string;
+// }
+
+import { Header } from '@/components/common/Header/Header';
 
 const ThemePage = () => {
-  return <Header />;
+  return <Header label="/login" title="선물하기" description="로그인" backgroundColor="#FFF" />;
 };
 
 export default ThemePage;
